Validation test for revision 2 of comp. phys. comm. paper. 1. Minor code reordering to make the code more like sac 2. An indexing error the the flux computation routines centdiff1 and centdiff2 was corrected
We have modified the way boundaries are updated making the code more like SAC. We employ a loop 
which applies boundary conditions for each, field, direction and boundary (i.e. top/bottom)

This has been applied to the ot test and will need to be corrected for each of the tests.